Split up Analytics and Engagement code
