+++ src/components/counter/FavCounter.tsx
@@ -9,7 +9,7 @@ import Config from "#/constants/Config";
 import { Achievements } from "#/helpers/Achievements";
 import type { ShareableType } from "#/helpers/Sharing";
 import FavoritesStore from "#/helpers/Stores/FavoritesStore";
-import { getFavs, registerFav } from "#/helpers/network/Analytics";
+import { getFavs, registerFav } from "#/helpers/network/Engagement";
 import { updateBadgeState } from "#/helpers/provider/BadgeProvider";
 import { useCorporateColor } from "#/hooks/useAppColorScheme";
 import type { FaveableType } from "#/types";
