+++ src/screens/Home/components/article/Article.tsx
@@ -16,7 +16,8 @@ import { onLinkPress } from "#/helpers/Linking";
 import { onShare } from "#/helpers/Sharing";
 import Statistics from "#/helpers/Statistics";
 import PersonalStore from "#/helpers/Stores/PersonalStore";
-import { registerEvent, registerViews } from "#/helpers/network/Analytics";
+import { registerEvent } from "#/helpers/network/Analytics";
+import { registerViews } from "#/helpers/network/Engagement";
 import { stripVisualComposerShortcodes } from "#/helpers/utils/posts";
 import { useAppColorScheme } from "#/hooks/useAppColorScheme";
 import type { ArticleProperties, HttpsUrl } from "#/types";
